allow to reuse .d.ts from external npm modules (which are plugins) rename the module to be @theia/plugin allow to do the start/stop of an extension automatically after loading the plugin
