+++ packages/plugin/src/index.js
@@ -1,0 +1,10 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+    value: true
+});
+
+var empty = {}
+
+exports.default = empty;
+module.exports = exports['default'];
