+++ packages/plugin/src/plugin/command-registry.ts
@@ -10,7 +10,7 @@
  */
 import { CommandRegistryExt, PLUGIN_RPC_CONTEXT as Ext, CommandRegistryMain } from '../api/plugin-api';
 import { RPCProtocol } from '../api/rpc-protocol';
-import * as theia from 'theia';
+import * as theia from '@theia/plugin';
 import { Disposable } from './types-impl';
 
 export type Handler = <T>(...args: any[]) => T | PromiseLike<T>;
